+++ src/api/theme/index.ts
@@ -1,0 +1,7 @@
+export { getThemes, getThemeInfo, getThemeProducts } from './apis';
+export type { ThemeProductsResponse } from './apis';
+export {
+  useThemesQuery,
+  useThemeInfoQuery,
+  useThemeProductsQuery,
+} from './query';
